+++ imports/ui/PrivateRoute.jsx
@@ -5,7 +5,7 @@ import { Redirect, Route } from 'react-router-dom'
 const PrivateRoute = ({ component: Component, ...rest }) => {
 
     // Add your own authentication on the below line.
-    const isLoggedIn = Meteor.user();
+    const isLoggedIn = !Meteor.user();
 
     return (
         <Route
