+++ package.json
@@ -15,6 +15,7 @@
     "react-dom": "^16.8.6",
     "react-router": "^5.0.0",
     "react-router-dom": "^5.0.0",
+    "react-star-ratings": "^2.3.0",
     "simpl-schema": "^1.5.5",
     "styled-components": "^4.2.0"
   },
